+++ src/includes/classes/App.php
@@ -183,7 +183,7 @@ class App extends Classes\Core\Base\Core
      *
      * @var string Version.
      */
-    const VERSION = '161009.27704'; //v//
+    const VERSION = '161009.47572'; //v//
 
     /**
      * Constructor.